! Switch to gauge for JVM memory metrics
README.md states that JVM memory metrics are recorded using a gauge, but
the actual implementation uses a histogram. A histogram is not suitable
for this kind of metric, as the distribution of allocated memory is
typically of little interest, and long-term average memory consumption is
not a particular meaningful metric.

By switching to gauges for JVM memory metrics, it will be possible to
read the latest memory usage of an application which is more useful
for monitoring.